+++ src/js/defaults.js
@@ -1,6 +1,0 @@
-export default {
-	format: 'markdown',
-	ignorePinned: false,
-	protocolBlacklist: [ 'chrome-devtools:', 'chrome:', 'chrome-extension:', 'opera:' ],
-	domainBlacklist: [ 'mail.google.com', 'inbox.google.com' ]
-};