+++ src/js/lib-browser/aceThemeList.js
@@ -1,0 +1,37 @@
+export var lightThemes = {
+	chrome:          'Chrome',
+	clouds:          'Clouds',
+	crimson_editor:  'Crimson Editor',
+	dawn:            'Dawn',
+	dreamweaver:     'Dreamweaver',
+	eclipse:         'Eclipse',
+	github:          'GitHub',
+	solarized_light: 'Solarized Light',
+	textmate:        'TextMate',
+	tomorrow:        'Tomorrow',
+	xcode:           'XCode',
+	kuroir:          'Kuroir',
+	katzenmilch:     'KatzenMilch'
+};
+
+export var darkThemes = {
+	ambiance: 'Ambiance',
+	chaos: 'Chaos',
+	clouds_midnight: 'Clouds Midnight',
+	cobalt: 'Cobalt',
+	idle_fingers: 'idle Fingers',
+	kr_theme: 'krTheme',
+	merbivore: 'Merbivore',
+	merbivore_soft: 'Merbivore Soft',
+	mono_industrial: 'Mono Industrial',
+	monokai: 'Monokai',
+	pastel_on_dark: 'Pastel on dark',
+	solarized_dark: 'Solarized Dark',
+	terminal: 'Terminal',
+	tomorrow_night: 'Tomorrow Night',
+	tomorrow_night_blue: 'Tomorrow Night Blue',
+	tomorrow_night_bright: 'Tomorrow Night Bright',
+	tomorrow_night_eighties: 'Tomorrow Night 80s',
+	twilight: 'Twilight',
+	vibrant_ink: 'Vibrant Ink'
+};